Coded the second layer of obfuscation
What and Why:
Trying to add a second layer of obfuscation to the input messages. This is so that messages will have gibberish on both ends.
How and consequences:
Added simple phrases to the beginning and end of the message.
I believe this may actually weaken encryption that I do later. So in the future maybe randomize the words and number of words and position.
For now this is just a practice for writing a pure Facade design pattern.
A pattern that decouples complex interaction of classes by supplying a simple interface, while still allowing access to the complex subcomponents.